added check for qrcode before deletion

+++ builds/signals.py
@@ -18,4 +18,5 @@ def delete_assoc_cateogries(sender, instance, **kwargs):
     if instance.interior_parts.count() > 0:
         for item in instance.interior_parts.all():
             item.delete()
-    os.remove(os.path.join(settings.MEDIA_ROOT, instance.qrcode.name))
+    if instance.qrcode:
+        os.remove(os.path.join(settings.MEDIA_ROOT, instance.qrcode.name))
